+++ backend/uclapi/dashboard/templates/dashboard.html
@@ -7,9 +7,6 @@
   <title>UCL API Dashboard</title>
 
   <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1.0" />
-  <link rel="stylesheet" href="https://unpkg.com/purecss@0.6.2/build/pure-min.css"
-    integrity="sha384-UQiGfs9ICog+LwheBSRCt1o5cbyKIHbwjWscjemyBMT9YCUMZffs6UqUTd0hObXD" crossorigin="anonymous">
-  <link rel="stylesheet" href="https://unpkg.com/purecss@0.6.2/build/grids-responsive-min.css">
   <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
   {% render_bundle 'dashboard' 'css' %}
 </head>
@@ -17,7 +14,6 @@
 
 <body>
   <div class="app"></div>
-  <script src="https://use.fontawesome.com/c70d56d6fc.js"></script>
   {% render_bundle 'vendors' 'js' %}
   {% render_bundle 'dashboard' 'js' %}
 </body>
