+++ frontend/package.json
@@ -44,12 +44,8 @@
   },
   "dependencies": {
     "@material-ui/core": "^4.11.0",
-    "ajv": "^6.12.5",
-    "aphrodite": "^2.4.0",
     "axios": "^0.20.0",
     "dayjs": "^1.8.36",
-    "glamor": "^2.20.40",
-    "immutability-helper": "^3.1.1",
     "include-media": "^1.4.9",
     "js-cookie": "^2.2.1",
     "prop-types": "^15.7.2",
@@ -65,7 +61,6 @@
     "react-router-dom": "^5.2.0",
     "react-syntax-highlighter": "^13.5.3",
     "react-tabs": "^3.1.1",
-    "react-transition-group": "^4.4.1",
     "whatwg-fetch": "^3.4.1"
   },
   "devDependencies": {
@@ -81,7 +76,6 @@
     "babel-loader": "^8.1.0",
     "babel-plugin-transform-imports": "^2.0.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "babelify": "^10.0.0",
     "css-loader": "^4.3.0",
     "dotenv": "^8.2.0",
     "eslint": "^7.9.0",
