+++ frontend/src/pages/Authorise/Authorise.scss
@@ -1,0 +1,73 @@
+@import "../../sass/core.scss";
+@import '../../sass/helpers';
+
+.authorise-page {
+  font-family: "Roboto", "Calibri", Helvetica, sans-serif;
+
+  .centered {
+    text-align: center;
+  }
+
+  .navbar {
+    @include box-shadow(1);
+
+    align-items: center;
+    background-color: white;
+    border-bottom: 1px solid rgba(0, 0, 0, 0.14);
+    display: flex;
+    height: $navbar-height;
+    position: fixed;
+    top: 0;
+    width: 100%;
+
+    img {
+      height: $navbar-height - 15px;
+      margin-left: 25px;
+    }
+
+    .logo-text {
+      display: inline;
+      font-size: 2em;
+      margin: 10px;
+    }
+  }
+
+  > .content {
+    margin: 0 auto;
+    margin-top: calc(#{$navbar-height} + 1rem);
+    max-width: 700px;
+    border: 1px solid black;
+    padding: 1rem 2rem;
+
+    ul {
+      li {
+        line-height: 150%;
+      }
+    }
+
+    hr {
+      margin: 1rem 0;
+    }
+
+    > .authorise-buttons {
+      display: flex;
+      flex-direction: row;
+
+      button {
+        border-radius: 2px;
+        display: inline-block;
+        padding: 10px 20px;
+        text-decoration: none;
+        width: fit-content;
+        cursor: pointer;
+        border: none;
+
+        &.authorise-button {
+          background-color: $color-grey-dark;
+          color: $color-white;
+          margin-right: 1rem;
+        }
+      }
+    }
+  }
+}
