+++ frontend/src/pages/Authorise/index.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import Authorise from './Authorise.jsx'
+
+ReactDOM.render(
+  <Authorise />,
+  document.querySelector(`.app`)
+)
