+++ frontend/src/pages/Dashboard/index.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import Dashboard from './Dashboard.jsx'
+
+ReactDOM.render(
+  <Dashboard />,
+  document.querySelector(`.app`)
+)
