+++ frontend/webpack.prod.js
@@ -89,9 +89,9 @@ module.exports = {
     index: entryPointsPathPrefix + `/Home`,
     documentation: entryPointsPathPrefix + `/Documentation.jsx`,
     about: entryPointsPathPrefix + `/AboutPage.jsx`,
-    dashboard: entryPointsPathPrefix + `/Dashboard.jsx`,
+    dashboard: entryPointsPathPrefix + `/Dashboard`,
     marketplace: entryPointsPathPrefix + `/Marketplace.jsx`,
-    authorise: entryPointsPathPrefix + `/Authorise.jsx`,
+    authorise: entryPointsPathPrefix + `/Authorise`,
     warning: entryPointsPathPrefix + `/Warning.jsx`,
     404: entryPointsPathPrefix + `/404.jsx`,
     500: entryPointsPathPrefix + `/500.jsx`,
